// packages/sdk-ts/src/schemas/generated.ts
export interface paths {
// ...
    "/fbl/fulfillment_sku_relation/write": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** BuildFulfillmentSkuRelation */
        post: operations["postFblFulfillmentskurelationWrite"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/fulfillment_order/cancel": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CancelFulfillmentOrderForMCL */
        post: operations["postFblFulfillmentorderCancel"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/inbound_reservation/cancel": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CancelInboundReservation */
        post: operations["postFblInboundreservationCancel"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/outbound_order/cancel": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CancelOutboundOrder */
        post: operations["postFblOutboundorderCancel"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/vas/cancelVasOrder": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** CancelVasOrder4FBL */
        get: operations["getFblVasCancelVasOrder"];
        put?: never;
        /** CancelVasOrder4FBL */
        post: operations["postFblVasCancelVasOrder"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/inbound_order/cancel": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CancelnBoundOrder */
        post: operations["postFblInboundorderCancel"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/inbound_reservation/check": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** CheckInboundReservationSlot */
        get: operations["getFblInboundreservationCheck"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/fulfillment_order/create": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CreateFulfillmentOrderForMCL */
        post: operations["postFblFulfillmentorderCreate"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/fulfillment_order_pnf/create": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CreateFulfillmentOrderForMCLV2PNF */
        post: operations["postFblFulfillmentorderpnfCreate"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/fulfillment_sku/create": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CreateFulfillmentSkuDecouple */
        post: operations["postFblFulfillmentskuCreate"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/fulfillment_sku_fbl/create": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CreateFulfillmentSkuForFBL */
        post: operations["postFblFulfillmentskufblCreate"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/inbound_order/create": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CreateInboundOrder */
        post: operations["postFblInboundorderCreate"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/inbound_reservation/create": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CreateInboundReservation */
        post: operations["postFblInboundreservationCreate"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/outbound_order/create": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CreateOutBoundOrder */
        post: operations["postFblOutboundorderCreate"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/product_reinbound/create": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** CreateProductReinboundOrderForMCL */
        post: operations["postFblProductreinboundCreate"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/vas/createVasOrder": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** CreateVasOrder4FBL */
        get: operations["getFblVasCreateVasOrder"];
        put?: never;
        /** CreateVasOrder4FBL */
        post: operations["postFblVasCreateVasOrder"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/channel_stocks/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** GetChannelStocksForMCL */
        get: operations["getFblChannelstocksGet"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/fulfillment_products/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** GetFulfillmentProductDetail */
        get: operations["getFblFulfillmentproductsGet"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/fulfillment_sku_list/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** GetFulfillmentSkuListForMCL */
        get: operations["getFblFulfillmentskulistGet"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/fulfillment_sku_relation/get_by_sc_item": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** GetFulfillmentSkuRelationByScItem */
        get: operations["getFblFulfillmentskurelationGetbyscitem"];
        put?: never;
        /** GetFulfillmentSkuRelationByScItem */
        post: operations["postFblFulfillmentskurelationGetbyscitem"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/fulfillment_sku_relation/get_by_sku": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** GetFulfillmentSkuRelationBySku */
        get: operations["getFblFulfillmentskurelationGetbysku"];
        put?: never;
        /** GetFulfillmentSkuRelationBySku */
        post: operations["postFblFulfillmentskurelationGetbysku"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/fulfillment_sku_relation/get_by_sc_items": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** GetFulfillmentSkuRelationsByScItems */
        get: operations["getFblFulfillmentskurelationGetbyscitems"];
        put?: never;
        /** GetFulfillmentSkuRelationsByScItems */
        post: operations["postFblFulfillmentskurelationGetbyscitems"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/fulfillment_sku_relation/get_by_skus": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** GetFulfillmentSkuRelationsBySkus */
        get: operations["getFblFulfillmentskurelationGetbyskus"];
        put?: never;
        /** GetFulfillmentSkuRelationsBySkus */
        post: operations["postFblFulfillmentskurelationGetbyskus"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/icp_order/file": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** GetIcpOrderFile */
        get: operations["getFblIcporderFile"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/fbl/inbound_order_detail/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** GetInboundOrderDetail */
        get: operations["getFblInboundorderdetailGet"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}